feat: add support for CSP (#72)
This commit adds support for running with CSP. CSP can be enabled
on a per page basis. If this is done, fresh will set a default CSP that
is very restrictive. Users can modify the CSP with the `useCSP` hook.
Fresh will add its own script resources to the CSP automatically.