prefix asset urls with a `X-Fresh-Base-Url` header if present.
In order to "mount" a Fresh website onto a path at another website,
the subpath of that website needs to be prepended to all assets it
serves.

For example, if I want to mount my fresh site at
`https://example.com/my-fresh`, then whenever Fresh loads the
sourcecode of an island into the DOM, instead of using
`/_frsh/.../island-my-island.js` as the `<script src>` of an island,
it should use `/my-fresh/_frsh/.../island-my-island.js` or more fully:
`https://example.com/my-fresh/_frsh/.../island-my-island.js`

This reads a header `X-Fresh-Base-Url`, and if present, prepends it to
any asset urls. This allows the mounting site to instruct frsh to act
as though it were living at the subpath.