chore: remove `www` workflow (#1282)
This change removes the `www` workflow, simplifying CI to one workflow.
This is made more accessible, thanks to #1281. This also fixes PR
merging nits for when the affected files lie only within the `/www`
folder. E.g.
https://github.com/denoland/fresh/pull/1259#issuecomment-1586050738

The same checks in the `www` workflow are performed in the `ci`
workflow.