feat: improve Hero responsiveness
The Hero component doesn't respond well to smaller display view width. This change aims to improve that behavior by forcing a minimum height and enforcing a minimum vertical padding. Additionally, the buttons will start as a column and will extend to a row on device width of `md` and bigger, just like the other components in the Fresh Components page.