fix: flaky `doesn't leak data across renderers` test (#1901)
Using `createHandler()` is simpler than starting a server and fetching a
page. `createHandler()` avoids unnecessary network requests, which can
be flaky when you only need to analyse parts of the returned request. I
think we should push for more use of `createHandler()` and only start a
server as a last resort.

This closes #1896, or rather the particular test mentioned there, but
other tests involving TCP connections can still occur. Again, their
remedy is to use `createHandler()` instead of starting a server.