docs: Update documentation for active link with Tailwind (#2417)
**Proposed change:**
- Document styling for ﻿aria-current attribute separately for Tailwind
and Twind.

**Disclaimer:**
I am new to open-source contributions and are unfamiliar with the
process for submitting a pull request. I have found some more areas of
improvement in the Fresh documentation. Should this PR not be
up-to-standards on how to contribute, please guide me in the right
direction. I have a few more suggestions on issues I found while
developing a new Fresh app.

Any guidance would be helpful, either reach out to me on GitHub, Discord
(bastianplsfix) or X (bastianplsfix).

Best,
Sebastian 🦕