feat: allow `ctx.render()` to skip wrappers (#3170)
Allow `ctx.render()` to bypass app wrapper and layouts.

- I originally intended this as a solution for middleware (`app.use()`)
to have the same capability as routes (`app.route()`) to solve the
`devErrorOverlay()` including wrappers.
- The docs mention it's possible to "opt-out" during `ctx.render()`
> Every
[ctx.render()](https://fresh.deno.dev/docs/canary/concepts/context#render-1)
call will include the app wrapper component by default, unless opted
out.
  > \- https://fresh.deno.dev/docs/canary/advanced/app-wrapper

I don't need this myself currently, but it seemed a bit weird that
middleware was not able to bypass app wrapper or layouts, because it's
not possible to pass a `RouteConfig`. If this API is not something you
want to support, I could remove the code changes in favor of just
updating docs (above) and adding the tests.

### Open questions

Should it not be possible for "index" to use a layout? E.g. calls to
`/foo` should include the layout:

```ts
const app = new App()
  .layout("/foo", ({ Component }) => (<>foo/<Component /></>))
  .get("/foo", (ctx) => ctx.render(<>index</>));

// Should this not be "foo/index"? Currently renders "index"
```

Currently, I ignored the test for this, since it was not what I
expected. I tried to simulate the following folder layout, but with
imperative code:

```
<root>/routes
└── /foo
    ├── _layout.tsx
    └── index.tsx
```