docs: move CoC to .github folder
Helps reduce the amount of files in the root files on the root, as well
making it more consistent placing it next to CONTRIBUTING.md.

According to [GitHub docs](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-code-of-conduct-to-your-project#adding-a-code-of-conduct-manually),
the `.github/` folder is one of the recommended places, and should be
picked up by GitHub for links and on the project homepage.